+++ pkg/agent/drbd.go
@@ -17,14 +17,15 @@ type DrbdResources interface {
 	// StartUpdates starts the process of updating the current state of DRBD resources.
 	StartUpdates(ctx context.Context) error
 	// Get returns the resource state at the time the last update was made.
-	Get() []DrbdResourceState
+	Get() map[string]*DrbdResource
 }
 
 type drbdResources struct {
 	interval time.Duration
 
-	lock      sync.Mutex
-	resources []DrbdResourceState
+	lock    sync.Mutex
+	states  []DrbdResourceState
+	configs []DrbdConfiguration
 }
 
 func NewDrbdResources(resync time.Duration) DrbdResources {
@@ -38,13 +39,14 @@ func (d *drbdResources) StartUpdates(ctx context.Context) error {
 	for {
 		klog.V(3).Info("updating drbd state")
 
-		current, err := execDrbdSetup(ctx)
+		configs, states, err := execDrbdSetup(ctx)
 		if err != nil {
 			return err
 		}
 
 		d.lock.Lock()
-		d.resources = current
+		d.configs = configs
+		d.states = states
 		d.lock.Unlock()
 
 		select {
@@ -56,7 +58,7 @@ func (d *drbdResources) StartUpdates(ctx context.Context) error {
 	}
 }
 
-func execDrbdSetup(ctx context.Context) ([]DrbdResourceState, error) {
+func execDrbdSetup(ctx context.Context) ([]DrbdConfiguration, []DrbdResourceState, error) {
 	klog.V(4).Infof("Checking if DRBD is loaded")
 
 	// We check for the presence of DRBD here. We don't care about the specific DRBD version, for those cases
@@ -64,35 +66,70 @@ func execDrbdSetup(ctx context.Context) ([]DrbdResourceState, error) {
 	_, err := os.Stat("/proc/drbd")
 	if os.IsNotExist(err) {
 		klog.V(3).Infof("DRBD not ready, no resources to monitor")
-		return nil, nil
+		return nil, nil, nil
 	}
 
 	klog.V(4).Info("Command: drbdsetup status --json")
 
 	out, err := exec.CommandContext(ctx, "drbdsetup", "status", "--json").Output()
 	if err != nil {
-		return nil, fmt.Errorf("failed to execute drbdsetup status --json: %w", err)
+		return nil, nil, fmt.Errorf("failed to execute drbdsetup status --json: %w", err)
 	}
 
 	klog.V(5).Infof("Command result: %s", string(out))
 
-	var result []DrbdResourceState
-	err = json.Unmarshal(out, &result)
+	var states []DrbdResourceState
+	err = json.Unmarshal(out, &states)
 	if err != nil {
-		return nil, fmt.Errorf("failed to parse drbdsetup json: %w", err)
+		return nil, nil, fmt.Errorf("failed to parse drbdsetup status --json: %w", err)
 	}
 
-	return result, nil
+	klog.V(4).Info("Command: drbdsetup show --json")
+
+	out, err = exec.CommandContext(ctx, "drbdsetup", "show", "--json").Output()
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to execute drbdsetup show --json: %w", err)
+	}
+
+	klog.V(5).Infof("Command result: %s", string(out))
+
+	var configs []DrbdConfiguration
+	err = json.Unmarshal(out, &configs)
+	if err != nil {
+		return nil, nil, fmt.Errorf("failed to parse drbdsetup show --json: %w", err)
+	}
+
+	return configs, states, nil
 }
 
-func (d *drbdResources) Get() []DrbdResourceState {
+func (d *drbdResources) Get() map[string]*DrbdResource {
 	d.lock.Lock()
 	defer d.lock.Unlock()
 
-	result := make([]DrbdResourceState, len(d.resources))
-	copy(result, d.resources)
+	result := make(map[string]*DrbdResource)
+	for _, config := range d.configs {
+		result[config.Resource] = &DrbdResource{
+			Name:   config.Resource,
+			Config: config,
+		}
+	}
+
+	for _, state := range d.states {
+		if _, ok := result[state.Name]; !ok {
+			klog.V(4).Infof("Got DRBD state without config %s, skipping", state.Name)
+			continue
+		}
+
+		result[state.Name].State = state
+	}
 
 	return result
+}
+
+type DrbdResource struct {
+	Name   string
+	Config DrbdConfiguration
+	State  DrbdResourceState
 }
 
 type DrbdConnection struct {
@@ -147,3 +184,15 @@ func (d *DrbdResourceState) HasQuorum() bool {
 
 	return true
 }
+
+// DrbdConfiguration is the parsed output of "drbdsetup show --json".
+type DrbdConfiguration struct {
+	Resource string `json:"resource"`
+	Options  struct {
+		Quorum string `json:"quorum,omitempty"`
+	} `json:"options"`
+}
+
+const (
+	QuorumMajority = "majority"
+)