+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Automatically skip volumes that have no quorum policy enabled.
+
 ## [1.3.1] - 2025-10-22
 
 ### Changed