+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.3] - 2026-07-08
+
 ### Fixed
 
 - Do not force-demote a DRBD resource whose IO is suspended by explicit user action (for example while a
@@ -163,7 +165,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Deployment example
 - README with motivating example
 
-[Unreleased]: https://github.com/piraeusdatastore/piraeus-ha-controller/compare/v1.3.2...HEAD
+[Unreleased]: https://github.com/piraeusdatastore/piraeus-ha-controller/compare/v1.3.3...HEAD
+[1.3.3]: https://github.com/piraeusdatastore/piraeus-ha-controller/compare/v1.3.2...v1.3.3
 [1.3.2]: https://github.com/piraeusdatastore/piraeus-ha-controller/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/piraeusdatastore/piraeus-ha-controller/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/piraeusdatastore/piraeus-ha-controller/compare/v1.2.3...v1.3.0